requirements.txt
streamlit==1.32.0
pandas==2.1.3
plotly==5.18.0
numpy==1.26.2
python-dotenv==1.0.0
azure-storage-blob==12.19.0
pillow==10.1.0
scipy==1.11.4
pytz==2023.3
matplotlib==3.8.2